src/api/requirements_api.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
scikit-learn>=1.0.0
python-dotenv>=0.21.0
requests>=2.28.0
anthropic>=0.5.0
openai>=1.1.0 